refactor(tests): mirror the monorepo layout, one file per route/area
