+++ tests/package.json
@@ -1,14 +1,13 @@
 {
-  "name": "@tests/golden",
+  "name": "@tests/app",
   "version": "0.0.0",
   "private": true,
   "type": "module",
   "scripts": {
     "check-types": "tsc --noEmit",
     "test": "bun test --timeout 60000",
-    "test:api": "bun test --timeout 60000 specs/api",
-    "test:web": "bun test --timeout 60000 specs/web",
-    "test:e2e": "bun test --timeout 60000 specs/e2e"
+    "test:api": "bun test --timeout 60000 api",
+    "test:web": "bun test --timeout 60000 web"
   },
   "devDependencies": {
     "@packages/config": "workspace:*",
