+++ web/next/content/docs/getting-started/project-structure.mdx
@@ -21,7 +21,7 @@ Three import aliases wire it together: `@api/hono` gives the frontend its API ty
 │  ├─ db/         # Drizzle schema + migrations
 │  ├─ env/        # type-safe, validated environment variables
 │  └─ config/     # TS base, tsdown factory, and site.ts (the brand)
-└─ tests/golden/  # black-box golden suite: every endpoint, route, interaction
+└─ tests/         # black-box app test suite, mirroring api/ and web/: every endpoint, route, interaction
 ```
 
 ## `api/hono`
